+++ src/main/java/com/example/voucher_manager/domain/voucher/VoucherRestController.java
@@ -4,8 +4,6 @@ import com.example.voucher_manager.web.Util.Status;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
@@ -14,18 +12,18 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
-@Controller
-public class VoucherController {
+@RestController
+public class VoucherRestController {
 
     private final VoucherService voucherService;
 
-    private static final Logger logger = LoggerFactory.getLogger(VoucherController.class);
+    private static final Logger logger = LoggerFactory.getLogger(VoucherViewController.class);
 
     private final Status RESPONSE_SUCCESS = Status.SUCCESS;
 
     DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS");
 
-    public VoucherController(VoucherService voucherService) {
+    public VoucherRestController(VoucherService voucherService) {
         this.voucherService = voucherService;
     }
 
@@ -82,55 +80,5 @@ public class VoucherController {
     public ResponseEntity<Status> deleteVoucher(@PathVariable("voucherId") UUID voucherId) {
         var response = voucherService.deleteVoucher(voucherId);
         return response ? ResponseEntity.ok(RESPONSE_SUCCESS) : ResponseEntity.internalServerError().build();
-    }
-
-    // ===== VIEWS =====
-    @GetMapping("/vouchers")
-    public String viewVouchersPage(Model model) {
-        List<VoucherDto> convertObject = voucherService.findAll().stream()
-                .map(VoucherDto::from)
-                .toList();
-        model.addAttribute("vouchers", convertObject);
-        return "views/voucher/vouchers";
-    }
-
-    @GetMapping("/vouchers/{voucherId}")
-    public String findVoucher(@PathVariable UUID voucherId, Model model) {
-        var findVoucher = voucherService.findVoucher(voucherId);
-        if (findVoucher.isPresent()) {
-            model.addAttribute("voucher", VoucherDto.from(findVoucher.get()));
-            return "views/voucher/voucher-details";
-        }
-        return "views/error/404";
-    }
-
-    @PostMapping("/vouchers/{voucherId}")
-    public String modifyVoucher(VoucherDto voucherDto) {
-        logger.info("Got voucher data modify request {}", voucherDto);
-        voucherService.updateVoucher(VoucherDto.toEntity(voucherDto));
-        return "redirect:/vouchers";
-    }
-
-    @GetMapping("/vouchers/new")
-    public String viewCreateVoucherPage() {
-        return "views/voucher/voucher-new";
-    }
-
-    @PostMapping("/vouchers/new")
-    public String createVoucherPage(CreateVoucherRequest createVoucherRequest) {
-        logger.info("Got create voucher request {}", createVoucherRequest);
-        voucherService.createVoucher(
-                VoucherType.of(createVoucherRequest.voucherType()),
-                createVoucherRequest.discountInfo());
-        return "redirect:/vouchers";
-    }
-
-    @GetMapping("/vouchers/remove/{voucherId}")
-    public String removeVoucher(@PathVariable UUID voucherId) {
-        logger.info("Got voucher remove request {}", voucherId);
-        if (voucherService.deleteVoucher(voucherId)) {
-            return "redirect:/vouchers";
-        }
-        return "views/error/504";
     }
 }
